Rewriting node module
